fixed ui browse tab, segregated UI componenets in frontend/scr/componenets
